[type-stage] extract pure document-tree derivations into document-nodes-meta
Phase 2 of the document-tree walker/renderer separation — the
delegation landing
(thoughts/plans/2026-09-05-2159-document-tree-walker-renderer-separation.md).

The renderer-independent halves move to the new DOM-free module
document-nodes-meta.mjs: the spec helpers, the element
rendering-plan derivation (UIDocumentElement constructor seam), the
attr-driven DOM diff, next-sibling typeSpec resolution, styler
dependency-mapping construction, and the mark-descriptor pipeline.
The viewer keeps its class shapes and call-graph direction and
delegates — a pure code move, behavior-neutral.

Not moved (Phase 3 scope): the UIDocumentNodes slot-insertion quartet
and the insertDocumentNode monkey-patch — their split lands with the
renderer-attachment interface.

Full suite green (255 tests).

model: moonshotai/kimi-k3
provider: openrouter
agent: goose v1.46.0
